properly update dynamicBindingList when value is an array
